Enabled distributed testing suite.
Two changes. To the existing, intended distributed test, added a `sleep` command and `stop-workers` so that the process could reach completion (prior to, the workers were simply left standing, awaiting additional tasks, so the test never 'completed' and so hung). The other, more important addition, was of another test, remote_feature_demo, which is simply the feature_demo with an additional step in the study, 'stop-workers', which 'exits' at the end of the workflow for a more graceful exit than the previous change.