Enabled distributed testing suite. (#333)
* Enabled distributed testing suite.

Many changes. This brings the distributed testing online and adds them to the CI, adding Makefile targets for both local and distributed tests. The previous distributed test hung forever because the workers were simply left standing, awaiting additional tasks, so the test never 'completed'. Remote_feature_demo fixes this - it is very nearly the feature_demo workflow, with an additional step in the study, 'stop-workers', which 'exits' at the end of the workflow. This required standing up a docker image of redis to run the distributed tests, so that change has been made in the CI.

Embedded in and commented out is another test, which adds a `sleep` command and `stop-workers` so that the process could more gracefully reach completion () rather than relying on the exit command, which shuts down all workers. This is unplugged this way as `stop-workers` is not currently functional, but, when that is online, this should be activated as well.

In a similar vein, there is embedded code for running against rabbitmq - this is not currently configured, but should be in the future for total test coverage of all supported configurations.

Lastly, upgraded the merlin config args to utilize the `test` arg. This arg had been in the source, but was unused (it read through the option and performed the exact same operations regardless). It now is invoked when running the distributed tests to examine the app_test.yaml.

Co-authored-by: Alexander Cameron Winter <winter27@quartz1148.llnl.gov>