adding old changes (#337)
* Update to the example optimization workflow. 

Originally merged in an incomplete state, this updates the example example optimization workflow. It renames the workflow to template_optimization, integrates the merlin spellbook to some of the scripts (not all that are possible), adds some test functions for examination, implements Jinja to develop the environment. This update will be released with 1.8.2 although additional work is desired for a future release - these future changes are recorded in github https://github.com/LLNL/merlin/pull/337.